rag-service/requirements.txt
python==3.11.*
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.10.*
pydantic-settings==2.7.*
httpx==0.28.*
structlog==24.4.*
prometheus-client==0.21.*
redis==5.2.*
asyncpg==0.30.*
sqlalchemy[asyncio]==2.0.*
pymilvus==2.4.*
minio==7.2.*
langfuse==2.56.*
python-multipart==0.0.18
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*

# RAG-specific dependencies
llama-index==0.10.*
llama-index-vector-stores-milvus==0.1.*
sentence-transformers==3.0.*
pillow==10.1.*
PyPDF2==4.0.*
python-docx==0.8.*
beautifulsoup4==4.12.*
lxml==4.9.*
markdownify==0.11.*
marker-pdf==0.3.*
torch==2.2.*
torchvision==0.17.*

# Utility
python-dotenv==1.0.*
